Separate SMT dialect extension
