+++ lighthouse/dialects/__init__.py
@@ -1,8 +1,10 @@
 def register_and_load(**kwargs):
+    from . import smt_ext
     from .transform import transform_ext
-    from .transform import smt_ext
+    from .transform import smt_ext as td_smt_ext
     from .transform import tune_ext
 
-    transform_ext.register_and_load(**kwargs)
     smt_ext.register_and_load(**kwargs)
+    transform_ext.register_and_load(**kwargs)
+    td_smt_ext.register_and_load(**kwargs)
     tune_ext.register_and_load(**kwargs)
