+++ lighthouse/dialects/smt_ext/__init__.py
@@ -1,0 +1,9 @@
+from .dialect import register_and_load
+from .dialect import SMTIntValue
+from .dialect import assert_
+
+__all__ = [
+    "SMTIntValue",
+    "assert_",
+    "register_and_load",
+]
