+++ lighthouse/dialects/transform/smt_ext/__init__.py
@@ -1,16 +1,12 @@
 from .dialect import register_and_load
 from .dialect import TransformSMTExtensionDialect
-from .dialect import SMTIntValue
-from .dialect import assert_
 
 from .ops.constrain_params import constrain_params
 from .ops.constrain_params import ConstrainParamsOp
 
 __all__ = [
     "ConstrainParamsOp",
-    "SMTIntValue",
     "TransformSMTExtensionDialect",
-    "assert_",
     "constrain_params",
     "register_and_load",
 ]
